services/netsrv-py/requirements.txt
redis==5.2.0
aioredis==2.0.1
paho-mqtt==2.1.0
aiohttp==3.10.0
pyyaml==6.0.2
python-dotenv==1.0.0
asyncio-mqtt==0.16.2
aiobotocore==2.15.0